z2_plus_defconfig: disable MMC_TEST
MMC host test driver

Development driver that performs a series of reads and writes
to a memory card in order to expose certain well known bugs
in host controllers. The tests are executed by writing to the
test file in debugfs under each card. Note that whatever is
on your card will be overwritten by these tests.

This driver is only of interest to those developing or
testing a host driver. Most people should say N here.

Signed-off-by: Alin Jerpelea <alin.jerpelea@sonymobile.com>