techpack: audio: q6voice: Force AVCS CVP version 1 on legacy SoC
Force the AVCS version to CVP_VERSION_1 for all of the AVCS
services on all legacy SoCs: these ones cannot execute the
Q6 Service Version command on their Q6 DSP because of older
firmware not supporting it.
We know that for these SoCs the version is always 1, hence
forcing it is totally safe.

The version has been forced on the following SoCs:
MSM8956, APQ8056, MSM8996, MSM8998, SDM630, SDM660.